# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Comics Library API","version":"1.0.0"},"paths":{"/v1/series":{"get":{"tags":["library"],"summary":"List Series","operationId":"list_series_v1_series_get","parameters":[{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":25,"title":"Page Size"}},{"name":"page_token","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Page Token"}},{"name":"publisher","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by publisher","title":"Publisher"},"description":"Filter by publisher"},{"name":"title_search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Substring filter for series title","title":"Title Search"},"description":"Substring filter for series title"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListSeriesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["library"],"summary":"Create Series","operationId":"create_series_v1_series_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSeriesRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Series"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/series/{series_id}":{"get":{"tags":["library"],"summary":"Get Series","operationId":"get_series_v1_series__series_id__get","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Series"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["library"],"summary":"Update Series","operationId":"update_series_v1_series__series_id__patch","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSeriesRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Series"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["library"],"summary":"Delete Series","operationId":"delete_series_v1_series__series_id__delete","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/series/{series_id}/issues":{"get":{"tags":["library"],"summary":"List Issues","operationId":"list_issues_v1_series__series_id__issues_get","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":25,"title":"Page Size"}},{"name":"page_token","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Page Token"}},{"name":"story_arc","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by story arc exact match","title":"Story Arc"},"description":"Filter by story arc exact match"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListIssuesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["library"],"summary":"Create Issue","operationId":"create_issue_v1_series__series_id__issues_post","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateIssueRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Issue"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/series/{series_id}/issues/{issue_id}":{"get":{"tags":["library"],"summary":"Get Issue","operationId":"get_issue_v1_series__series_id__issues__issue_id__get","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}},{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Issue"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["library"],"summary":"Update Issue","operationId":"update_issue_v1_series__series_id__issues__issue_id__patch","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}},{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateIssueRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Issue"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["library"],"summary":"Delete Issue","operationId":"delete_issue_v1_series__series_id__issues__issue_id__delete","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}},{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/issues/{issue_id}/copies":{"get":{"tags":["library"],"summary":"List Copies","operationId":"list_copies_v1_issues__issue_id__copies_get","parameters":[{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":25,"title":"Page Size"}},{"name":"page_token","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Page Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListCopiesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["library"],"summary":"Create Copy","operationId":"create_copy_v1_issues__issue_id__copies_post","parameters":[{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCopyRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Copy"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/issues/{issue_id}/copies/{copy_id}":{"get":{"tags":["library"],"summary":"Get Copy","operationId":"get_copy_v1_issues__issue_id__copies__copy_id__get","parameters":[{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}},{"name":"copy_id","in":"path","required":true,"schema":{"type":"integer","title":"Copy Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Copy"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["library"],"summary":"Update Copy","operationId":"update_copy_v1_issues__issue_id__copies__copy_id__patch","parameters":[{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}},{"name":"copy_id","in":"path","required":true,"schema":{"type":"integer","title":"Copy Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCopyRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Copy"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["library"],"summary":"Delete Copy","operationId":"delete_copy_v1_issues__issue_id__copies__copy_id__delete","parameters":[{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}},{"name":"copy_id","in":"path","required":true,"schema":{"type":"integer","title":"Copy Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/series/{series_id}/issues/{issue_id}/copies/{copy_id}/images":{"post":{"tags":["library"],"summary":"Upload Copy Image","operationId":"upload_copy_image_v1_series__series_id__issues__issue_id__copies__copy_id__images_post","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}},{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}},{"name":"copy_id","in":"path","required":true,"schema":{"type":"integer","title":"Copy Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_copy_image_v1_series__series_id__issues__issue_id__copies__copy_id__images_post"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageUploadJob"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["library"],"summary":"List Copy Images","operationId":"list_copy_images_v1_series__series_id__issues__issue_id__copies__copy_id__images_get","parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}},{"name":"issue_id","in":"path","required":true,"schema":{"type":"integer","title":"Issue Id"}},{"name":"copy_id","in":"path","required":true,"schema":{"type":"integer","title":"Copy Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListCopyImagesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/jobs/{job_id}":{"get":{"tags":["jobs"],"summary":"Get Job","operationId":"get_job_v1_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageUploadJob"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_copy_image_v1_series__series_id__issues__issue_id__copies__copy_id__images_post":{"properties":{"image_type":{"$ref":"#/components/schemas/ImageType"},"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["image_type","file"],"title":"Body_upload_copy_image_v1_series__series_id__issues__issue_id__copies__copy_id__images_post"},"ComicImage":{"properties":{"series_id":{"type":"integer","title":"Series Id"},"issue_id":{"type":"integer","title":"Issue Id"},"copy_id":{"type":"integer","title":"Copy Id"},"image_type":{"$ref":"#/components/schemas/ImageType"},"file_name":{"type":"string","title":"File Name"},"relative_path":{"type":"string","title":"Relative Path"}},"additionalProperties":false,"type":"object","required":["series_id","issue_id","copy_id","image_type","file_name","relative_path"],"title":"ComicImage"},"Copy":{"properties":{"clz_comic_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Clz Comic Id"},"custom_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Label"},"format":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Format"},"grade":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grade"},"grader_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grader Notes"},"grading_company":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grading Company"},"raw_slabbed":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Slabbed"},"signed_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Signed By"},"slab_cert_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Slab Cert Number"},"purchase_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purchase Date"},"purchase_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Purchase Price"},"purchase_store":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purchase Store"},"purchase_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Purchase Year"},"date_sold":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date Sold"},"price_sold":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Sold"},"sold_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Sold Year"},"my_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"My Value"},"covrprice_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Covrprice Value"},"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"age":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Age"},"barcode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Barcode"},"cover_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cover Price"},"page_quality":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Page Quality"},"key_flag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Flag"},"key_category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Category"},"key_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Reason"},"label_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Label Type"},"no_of_pages":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"No Of Pages"},"variant_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variant Description"},"copy_id":{"type":"integer","title":"Copy Id","description":"Primary key for a copy"},"issue_id":{"type":"integer","title":"Issue Id"}},"additionalProperties":false,"type":"object","required":["copy_id","issue_id"],"title":"Copy"},"CreateCopyRequest":{"properties":{"clz_comic_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Clz Comic Id"},"custom_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Label"},"format":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Format"},"grade":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grade"},"grader_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grader Notes"},"grading_company":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grading Company"},"raw_slabbed":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Slabbed"},"signed_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Signed By"},"slab_cert_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Slab Cert Number"},"purchase_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purchase Date"},"purchase_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Purchase Price"},"purchase_store":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purchase Store"},"purchase_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Purchase Year"},"date_sold":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date Sold"},"price_sold":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Sold"},"sold_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Sold Year"},"my_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"My Value"},"covrprice_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Covrprice Value"},"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"age":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Age"},"barcode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Barcode"},"cover_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cover Price"},"page_quality":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Page Quality"},"key_flag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Flag"},"key_category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Category"},"key_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Reason"},"label_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Label Type"},"no_of_pages":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"No Of Pages"},"variant_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variant Description"}},"additionalProperties":false,"type":"object","title":"CreateCopyRequest"},"CreateIssueRequest":{"properties":{"issue_nr":{"type":"string","title":"Issue Nr","description":"CLZ normalized issue number"},"variant":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variant","description":"Variant designation","default":""},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"subtitle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subtitle"},"full_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Title"},"cover_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover Date"},"cover_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Cover Year"},"story_arc":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Story Arc"}},"additionalProperties":false,"type":"object","required":["issue_nr"],"title":"CreateIssueRequest"},"CreateSeriesRequest":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title","description":"Series display title"},"publisher":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Publisher","description":"Publisher name"},"series_group":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Series Group","description":"Grouping label"},"age":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Age","description":"Age category from CLZ"},"series_id":{"type":"integer","title":"Series Id","description":"User supplied unique identifier"}},"additionalProperties":false,"type":"object","required":["series_id"],"title":"CreateSeriesRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageType":{"type":"string","enum":["front","back","spine","staples","interior_front_cover","interior_back_cover","misc"],"title":"ImageType"},"ImageUploadJob":{"properties":{"job_id":{"type":"string","title":"Job Id"},"series_id":{"type":"integer","title":"Series Id"},"issue_id":{"type":"integer","title":"Issue Id"},"copy_id":{"type":"integer","title":"Copy Id"},"image_type":{"$ref":"#/components/schemas/ImageType"},"status":{"$ref":"#/components/schemas/JobStatus"},"detail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Detail"},"result":{"anyOf":[{"$ref":"#/components/schemas/ComicImage"},{"type":"null"}]}},"additionalProperties":false,"type":"object","required":["job_id","series_id","issue_id","copy_id","image_type","status"],"title":"ImageUploadJob"},"Issue":{"properties":{"issue_nr":{"type":"string","title":"Issue Nr","description":"CLZ normalized issue number"},"variant":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variant","description":"Variant designation","default":""},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"subtitle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subtitle"},"full_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Title"},"cover_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover Date"},"cover_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Cover Year"},"story_arc":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Story Arc"},"issue_id":{"type":"integer","title":"Issue Id","description":"Primary key for an issue"},"series_id":{"type":"integer","title":"Series Id"}},"additionalProperties":false,"type":"object","required":["issue_nr","issue_id","series_id"],"title":"Issue"},"JobStatus":{"type":"string","enum":["pending","in_progress","completed","failed"],"title":"JobStatus"},"ListCopiesResponse":{"properties":{"next_page_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Page Token","description":"Opaque token clients pass to retrieve the next page."},"copies":{"items":{"$ref":"#/components/schemas/Copy"},"type":"array","title":"Copies"}},"additionalProperties":false,"type":"object","required":["copies"],"title":"ListCopiesResponse"},"ListCopyImagesResponse":{"properties":{"images":{"items":{"$ref":"#/components/schemas/ComicImage"},"type":"array","title":"Images"}},"additionalProperties":false,"type":"object","required":["images"],"title":"ListCopyImagesResponse"},"ListIssuesResponse":{"properties":{"next_page_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Page Token","description":"Opaque token clients pass to retrieve the next page."},"issues":{"items":{"$ref":"#/components/schemas/Issue"},"type":"array","title":"Issues"}},"additionalProperties":false,"type":"object","required":["issues"],"title":"ListIssuesResponse"},"ListSeriesResponse":{"properties":{"next_page_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Page Token","description":"Opaque token clients pass to retrieve the next page."},"series":{"items":{"$ref":"#/components/schemas/Series"},"type":"array","title":"Series"}},"additionalProperties":false,"type":"object","required":["series"],"title":"ListSeriesResponse"},"Series":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title","description":"Series display title"},"publisher":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Publisher","description":"Publisher name"},"series_group":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Series Group","description":"Grouping label"},"age":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Age","description":"Age category from CLZ"},"series_id":{"type":"integer","title":"Series Id","description":"Primary key for a series"}},"additionalProperties":false,"type":"object","required":["series_id"],"title":"Series"},"UpdateCopyRequest":{"properties":{"clz_comic_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Clz Comic Id"},"custom_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Label"},"format":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Format"},"grade":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grade"},"grader_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grader Notes"},"grading_company":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Grading Company"},"raw_slabbed":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Slabbed"},"signed_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Signed By"},"slab_cert_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Slab Cert Number"},"purchase_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purchase Date"},"purchase_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Purchase Price"},"purchase_store":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Purchase Store"},"purchase_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Purchase Year"},"date_sold":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date Sold"},"price_sold":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Sold"},"sold_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Sold Year"},"my_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"My Value"},"covrprice_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Covrprice Value"},"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"age":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Age"},"barcode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Barcode"},"cover_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cover Price"},"page_quality":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Page Quality"},"key_flag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Flag"},"key_category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Category"},"key_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Key Reason"},"label_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Label Type"},"no_of_pages":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"No Of Pages"},"variant_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variant Description"}},"additionalProperties":false,"type":"object","title":"UpdateCopyRequest"},"UpdateIssueRequest":{"properties":{"issue_nr":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Issue Nr"},"variant":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Variant"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"subtitle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subtitle"},"full_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Title"},"cover_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover Date"},"cover_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Cover Year"},"story_arc":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Story Arc"}},"additionalProperties":false,"type":"object","title":"UpdateIssueRequest"},"UpdateSeriesRequest":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title","description":"Series display title"},"publisher":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Publisher","description":"Publisher name"},"series_group":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Series Group","description":"Grouping label"},"age":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Age","description":"Age category from CLZ"}},"additionalProperties":false,"type":"object","title":"UpdateSeriesRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}